StructuredRNAStringSet for storing DotBracketAnnotation alongside nucleotide sequences

The StructuredXStringSet class can be used to store structure information alongside RNA sequences. The class behaves like the QualityScaledXStringSet classes.

Please note, that this does not check for validity regarding base pairing capabilities.

StructuredRNAStringSet(x, structure)

dotbracket(x)

dotbracket(x) <- value

# S4 method for StructuredXStringSet
dotbracket(x)

# S4 method for StructuredXStringSet
dotbracket(x) <- value

readStructuredRNAStringSet(
  filepath,
  nrec = -1L,
  skip = 0L,
  seek.first.rec = FALSE,
  use.names = TRUE
)

writeStructuredXStringSet(x, filepath, append = FALSE, compress = FALSE, ...)

# S4 method for StructuredXStringSet
getBasePairing(x, compress = TRUE, return.sequence = FALSE)

# S4 method for StructuredXStringSet
getLoopIndices(x, bracket.type, warn.type.drops = TRUE)

Arguments

x

For the Structured*StringSet constructors: Either a character vector, or an RNAString, RNAStringSet object. For writeStructuredXStringSet: A StructuredRNAStringSet derivative.

structure, value

A DotBracketStringSet

use.names, type, filepath, nrec, skip, seek.first.rec, append, ...

See DotBracketStringSet-io

compress

See getBasePairing or DotBracketStringSet-io

return.sequence

TRUE(default) or FALSE: Whether the sequence should be returned in the base column.

bracket.type

getLoopIndices: Which dot bracket annotation type should be converted into loop indices? Only usable, if more than one is present. (1L = '()', 2L = '<>', 3L = '[]', 4L = '{}')

warn.type.drops

See getLoopIndices

Value

a StructuredRNAStringSet object.

Details

the dotbracket function allows access to the included DotBracketStringSet.

Examples

str <- DotBracketStringSet("(())")
seq <- RNAStringSet("AGCU")
sdbs <- StructuredRNAStringSet(seq,str)